% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundance.R
\name{abundance}
\alias{abundance}
\title{abundance - Abundance calculating function}
\usage{
abundance(df, first.col = 2, plot = T, legh = 50, legl = 10, ylab = "Abundance", xlab = "Samples", col = c("gray0", "gray48", "gray84"), ncol = 3, ...)
}
\arguments{
\item{df}{A data frame containing uploaded user dataset (list of taxa in first column, followed by columns of abundances with sample names in a header).}

\item{first.col}{Numeric (integer), the first column of samples (to skip non-relevant columns). By default, this parameter is set to ‘2’.}

\item{plot}{Should a barplot for results of calculations be plotted? By default, the plot is rendered.}

\item{legh}{xx}

\item{legl}{xx}

\item{ylab}{The text for the y axis label. By default, the text is setted as ‘Abundance’.}

\item{xlab}{The text for the x axis label. By default, the text is setted as ‘Samples’.}

\item{col}{A vector that contains three components (named after the three colours). By default, the colours ‘gray0’, ‘gray48’ and ‘gray84’ is used.}

\item{lncol}{The number of columns in the legend. By default, it is set to ‘3’.}

\item{...}{xx}
}
\value{
A data frame consisting of a column of calculated abundance values (‘F1’, ‘F2’, ‘Other’, ‘Total’) with samples in the rows.
}
\description{
Calculates number of singletons and doubletons in the samples and total abundance of samples.
}
\examples{
For this function, you must have a uploaded data frame. Calculation of abundance for the ‘test_data’ data frame.

Abundance_table <- abundance(test_data, legh = 150, legl = 12, col = c("white", "darkgray", "darkcyan"), ncol = 3)

}
