% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brilloun,eco.R
\name{brillouin.eco}
\alias{brillouin.eco}
\title{brillouin.eco - Brillouin index and evenness for Brillouin index}
\usage{
brillouin.eco(x, first.col = 2, table = T, graph = T)
}
\arguments{
\item{x}{data}

\item{first.col}{integer, the first column for the analysis}

\item{table}{logical, if TRUE table will be creating (T) or uncreating (F)}

\item{graph}{logical, if TRUE graphical output will be drawn graph}
}
\description{
Calculation and graphical visualization of Brillouin index and evenness for Brillouin index
}
\details{
The function brillouin.eco provides the calculations of Brillouin index (HB), maximal diversity value (Hbmax) and evenness for Brillouin index for every sample in the uploaded dataset (x). With dataset is worked from the user's specified column (first.col=2). This function returns tabular (table=T) and graphical outputs (graph=T).
}
\examples{
brillouin.eco(x)
}
\references{
Krebs C.J., 2014: Ecological Methodology. 3rd ed. (in prep).
Pielou E.C., 1966: The measurement of diversity in different types of biological collections. Journal of Theoretical Biology 13: 131–144.
}
