% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brillouin.R
\name{brillouin}
\alias{brillouin}
\title{brillouin - Brillouin index and evenness}
\usage{
brillouin(x, plot = T, ylab = "Brillouin's index", xlab = "Samples", col = "black",...)
}
\arguments{
\item{x}{A data frame uploaded by the loadData function (important because of the first column setting) containing the uploaded user dataset (list of taxa in the first column, followed by abundance columns with sample names in a header).}

\item{plot}{Should a barplot for results of calculations be plotted? By default, the plot is rendered.}

\item{ylab}{The text for the y axis label in plot. By default, the text is setted as ‘Brillouin's index’.}

\item{xlab}{The text for the x axis label. By default, the text is setted as ‘Samples’.}

\item{col}{The colour used in the plot. By default, it is set as ‘gray’.}

\item{...}{Arguments to be passed to methods, such grapfical parameters. Many methods will acccept the following arguments.}
}
\value{
A data frame consisting of a column of calculated Brillouin calculations (‘HB’, ‘HBmax’, ‘Even’) with samples in the rows.
}
\description{
Calculates Brillouin index and evenness and displays the results of the calculation in graphical and tabular form.
}
\examples{
# Calculation of the Brillouin calculations for the 'test_data' and plotted results. The colour used is 'darkcyan'. Data frame of the results is stored in an object called 'brillouin_tab'.

brillouin_tab <- brillouin(test_data, col = "darkcyan")

}
