% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jacren.R
\name{jacren}
\alias{jacren}
\title{jacren - Jaccard and Renconen similarity index calculating function}
\usage{
jacren(df, plot = T, cex = 0.75)
}
\arguments{
\item{df}{A data frame containing uploaded user dataset (list of taxa in first column, followed by columns of abundances with sample names in a header).}

\item{plot}{Should a plot for results of calculations be plotted? By default, the plot is rendered.}

\item{cex}{Numeric. Size of the text in the plot cells.}
}
\value{
xxx
}
\description{
Calculation of Jaccard and Renconen similarity index
}
\examples{
jacren(test_data)
}
