% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{loadData}
\alias{loadData}
\title{read.eco - Uploading the dataset}
\usage{
read.eco(first.col=2, na2null=T, attrib=T)
}
\arguments{
\item{first.col}{the column from which to work with the data}

\item{na2null}{converting NA values to null}

\item{attrib}{creating attributes}

\item{type}{nastení např. clipboard}
}
\description{
Uploading the dataset, performing basic analyses for the loaded table, creating table attributes.
}
\details{
The function read.eco appeals to users to copy the dataset from the spreadsheet. With uploaded dataset is worked from the user's specified column (first.col=2). This function also automatically converts NA values to null (na2null=T) and writes out attributes of the dataset.
}
\examples{
x<-read.eco()

}
