% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MM.R
\name{MM}
\alias{MM}
\title{mm.eco - Margalef's and Menhinick's index}
\usage{
mm.eco(x, first.col=2, table=1, graph=1)
}
\arguments{
\item{x}{data}

\item{first.col}{the column from which to work with the data}

\item{table}{creating table}

\item{graph}{creating graph}
}
\description{
Calculation of species richness indices (Margalef's and Menhinick's index)
}
\details{
The function mm.eco provides the calculations of Margalef's (DMg) and Menhinick's (DMm) index for uploaded dataset (x). With dataset is worked from the user's specified column (first.col=2). This function returns tabular (table=T) and graphical outputs (graph=T). Users can set the outputs as:
1 = table and graph are made for both indices
2 = table and graph are made only for Margalef's index
3 = table and graph are made only for Menhinick's index
}
\examples{
mm.eco(x)
mm_index.eco(x, table = 2, graph = 2)
mm_index.eco(x, table = 3, graph = 3)
}
\references{
Clifford H.T. & Stephenson W., 1975: An introduction to numerical classification. Academic Press: New York, San Francisco, ISBN 978-0-12-176750-1.
Magurran A.E., 2013: Measuring Biological Diversity. Wiley-Blackwell: Malden, Oxford, Carlton, ISBN 978-1-118-68792-5.
Whittaker R., 1977: Evolution of species diversity in land communities. Evolutionary Biology 10: 1–67.
}
