% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/richness.R
\name{richness}
\alias{richness}
\title{richness - Species richness calculating function}
\usage{
richness(df, first.col = 2, graph = T, legh = 10, legl = 10, ylab = "Number of species", xlab = "Samples", col = c("gray0", "gray48", "gray84"), lncol = 3, ...)
}
\arguments{
\item{df}{A data frame containing uploaded user dataset (list of taxa in first column, followed by columns of abundances with sample names in a header).}

\item{first.col}{Numeric (integer), the first column of samples (to skip non-relevant columns). By default, this parameter is set to ‘2’.}

\item{plot}{Should a barplot for results of calculations be plotted? By default, the plot is rendered.}

\item{legh}{xxx}

\item{legl}{xxx}

\item{ylab}{The text for the y axis label. By default, the text is setted as ‘Number of species’.}

\item{xlab}{The text for the x axis label. By default, the text is setted as ‘Samples’.}

\item{col}{A vector that contains three components (named after the three colours). By default, the colours ‘gray0’, ‘gray48’ and ‘gray84’ is used.}

\item{lncol}{The number of columns in the legend. By default, it is set to ‘3’.}

\item{...}{xxx}
}
\value{
A data frame consisting of a column of calculated species richness values (‘F1’, ‘F2’, ‘Other’, ‘Total’) with samples in the rows.
}
\description{
Calculates number of singletons and doubletons in the samples and total abundance of samples.
}
\examples{
xx
}
