% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shannon.R
\name{shannon}
\alias{shannon}
\title{shannon - Shannon index and evenness calculating function}
\usage{
shannon(df, first.col = 2, table = "CE", plot = T,  log = "e", ylab = "Shannon index value", col = "black", ...)
}
\arguments{
\item{df}{A data frame containing uploaded user dataset (list of taxa in first column, followed by columns of abundances with sample names in a header).}

\item{first.col}{Numeric (integer), the first column of samples (to skip non-relevant columns). By default, this parameter is set to ‘2’.}

\item{table}{Table output type. In case the table is equal to ‘C’ the table will be created only for Shannon calculations (Shannon index, the lower limit of Shannon index and the highest limit of Shannon index), for table equal to ‘E’ the output will be the table for evenness of Shannon index. In the case of table equal to ‘CE’ a sheet of two table type elements will be created for both Shannon calculations and evenness.}

\item{plot}{Should a plot for results of calculations be plotted? By default, the plot is rendered.}

\item{log}{Logarithm base type. The possibilities are ‘e’, for a base equal to Euler's number and ‘2’, for a base equal to two.}

\item{ylab}{The text for the y axis label in plot. By default, the text is setted as ‘Shannon index value’.}

\item{col}{A vector that contains five components (named after the five colours). By default, the color is set as ‘gray’.}

\item{...}{xxx}
}
\value{
A data frame consisting of a column of calculated species richness values (‘F1’, ‘F2’, ‘Other’, ‘Total’) with samples in the rows.
}
\description{
Calculation of Shannon index and evenness
}
\examples{
# Popsat nastavení následujících argumentů¨

shannon(test_data, log = "2", table = "C")
}
