% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SN.R
\name{SN}
\alias{SN}
\title{sn.eco - Species richness and abundance}
\usage{
sn.eco(x, first.col = 2, table = T, graph = T)
}
\arguments{
\item{x}{data}

\item{first.col}{the column from which to work with the data}

\item{table}{creating (T) or uncreating (F) table}

\item{graph}{creating (T) or uncreating (F) graph}
}
\description{
Calculation of species richness and abundance.
}
\details{
The function sn.eco provides the calculations of species richness and abundance for uploaded dataset (x). With dataset is worked from the user's specified column (first.col=2). This function provides tabular (table=T) and graphical outputs (graph=T).
}
\examples{
sn.eco(x)

}
