% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundance.R
\name{abundance}
\alias{abundance}
\title{abundance - Abundance}
\usage{
abundance(x, plot = T, lehver = 50, leghor = 10 , ylab = "Abundance", xlab = "Samples", col = c("gray0", "gray48", "gray84"), ncol = 3, ...)
}
\arguments{
\item{x}{A data frame uploaded by the loadData function (important because of the first column setting) containing the uploaded user dataset (list of taxa in the first column, followed by abundance columns with sample names in a header).}

\item{plot}{Should a barplot for results of calculations be plotted? By default, the plot is rendered.}

\item{legver}{Numeric. Setting the vertical position of the legend.}

\item{leghor}{Numeric. Setting the horizontal position of the legend.}

\item{ylab}{The text for the y axis label. By default, the text is setted as ‘Abundance’.}

\item{xlab}{The text for the x axis label. By default, the text is setted as ‘Samples’.}

\item{col}{A vector that contains three components (named after the three colours). By default, the colours ‘gray0’, ‘gray48’ and ‘gray84’ is used.}

\item{lncol}{The number of columns in the legend. By default, it is set to ‘3’.}

\item{...}{Arguments to be passed to methods, such graphical parameters. Many methods will acccept the following arguments.}
}
\value{
A data frame consisting of a column of calculated abundance values (‘F1’ – abundance of singletons, ‘F2’ – abundance od doubletons, ‘Other’, ‘Total’) with samples in the rows.
}
\description{
Calculates the number of singletons and doubletons in the samples and the total abundance of the samples and displays the results of the calculation in graphical and tabular form.
}
\examples{
# Calculation of the abundance for the 'test_data' with the position of the legend vertical set to '150' and the position of the legend length set to '12'. The number of columns in the legend is set to '3' and the colour palette used is 'white', 'darkgray' and 'darkcyan'. The data frame containing the calculation results is stored in an object called 'abundance_tab'.

abundance_tab <- abundance(test_data, lehver = 150, leghor = 12, lncol = 3, col = c("white", "darkgray", "darkcyan"))

}
