% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brillouin.R
\name{brillouin}
\alias{brillouin}
\title{brillouin - Brillouin index and evenness calculating function}
\usage{
rillouin(df, first.col = 2, plot = T, ylab = "Brillouin's index", col = "gray",...)
}
\arguments{
\item{df}{A data frame containing uploaded user dataset (list of taxa in first column, followed by columns of abundances with sample names in a header).}

\item{first.col}{Numeric (integer), the first column of samples (to skip non-relevant columns). By default, this parameter is set to ‘2’.}

\item{plot}{Should a barplot for results of calculations be plotted? By default, the plot is rendered.}

\item{ylab}{The text for the y axis label in plot. By default, the text is setted as ‘Brillouin's index’.}

\item{col}{A vector that contains five components (named after the five colours). By default, the color is set as ‘gray’.}

\item{...}{xxx}
}
\value{
A data frame consisting of a column of calculated species richness values (‘F1’, ‘F2’, ‘Other’, ‘Total’) with samples in the rows.
}
\description{
Calculation of Brillouin index and evenness
}
\examples{
Popsat nastavení následujících argumentů
přidat vice volání
brillouin(test_data.1, col = "darkcyan")

}
