% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dominance.R
\name{dominance}
\alias{dominance}
\title{dominance - Dominance classes analysing function.}
\usage{
xxx
}
\arguments{
\item{df}{A data frame containing uploaded user dataset (list of taxa in first column, followed by columns of abundances with sample names in a header).}

\item{table}{xxx}

\item{plot}{xxx}

\item{legh}{xxx}

\item{legl}{xxx}

\item{xlab}{The text for the x axis label. By default, the text is setted as ‘Samples’.}

\item{col}{A vector that contains five components (named after the five colours). By default, the colours ‘gray0’, ‘gray48’, ‘gray78’, ‘gray84’ and ‘gray90’ is used.}

\item{lncol}{The number of columns in the legend. By default, it is set to ‘6’.}

\item{bty}{xxx}

\item{...}{xxx}

\item{ylabA}{xxx}

\item{ylabR}{xx}
}
\value{
xx
}
\description{
Tischler's dominance classes analyses
}
\examples{
xxx

}
