% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jacren.R
\name{jacren}
\alias{jacren}
\title{jacren - Jaccard and Renkonen similarity index}
\usage{
jacren(x, plot = T, cex = 0.4)
}
\arguments{
\item{x}{A data frame uploaded by the loadData function (important because of the first column setting) containing the uploaded user dataset (list of taxa in the first column, followed by abundance columns with sample names in a header).}

\item{plot}{Should a plot for results of calculations be plotted? By default, the plot is rendered.}

\item{cex}{Numeric. Size of the text in the plot cells. By default, it is set to 0.4.}
}
\value{
Numerical. The column and row headers contain the sample names, the cells contain the calculated index values.
}
\description{
Calculates Jaccard and Renkonen similarity index and displays the results of the calculation in graphical and tabular form.
}
\examples{
# Calculation of the Jaccard and Renconen similarity index for the 'test_data' and plotted results.  The tabular output is stored in an object called 'jacren_tab'.

jacren_tab <- jacren(test_data)

}
