% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{loadData}
\alias{loadData}
\title{Data uploading}
\usage{
loadData(file = "clipboard", first.col = 2, na2null = T, attrib = T)
}
\arguments{
\item{first.col}{Numeric (integer), the first column of samples (to identify non-relevant columns). By default, this parameter is set to ‘2’ (typically, the species names are usually found in the first column of the table). The following functions in this package work with data from this column.}

\item{na2null}{Logical, TRUE in case of converting NA values into null.}

\item{attrib}{Logical, TRUE in case of including attributes (‘names’, ‘class’, ‘row.names’, ‘Spec_col’, ‘Number of species’, ‘Number of samples’, ‘First column’).}
}
\value{
A data frame uploaded by the user containing a list of taxa in the first column, and abundance data in following columns, with sample names in the column header.
}
\description{
The function allows for simple loading of user data, checking the basic parameters of the data and creating the object (data frame) attributes.
}
\examples{
Upload the test data and store it in an object named test_data.

test_data <- loadData()

}
