% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{loadData}
\alias{loadData}
\title{loadData - Data uploading function}
\usage{
loadData(file = "clipboard", first.col = 2, na2null = T, attrib = T)
}
\arguments{
\item{first.col}{Numeric (integer), the first column of samples (to skip non-relevant columns). By default, this parameter is set to ‘2’.}

\item{na2null}{Logical, true in case of converting NA values into null.}

\item{attrib}{Logical, true in case of creating attributes.}
}
\value{
A data frame uploaded by the user containing a list of taxa in the first column, and abundance or presence/absence data in following columns, with sample names in the column header.
}
\description{
The function allows for simple loading of user data, checking the basic parameters of the data and creating the object (data frame) attributes.
}
\examples{
Uploading testing data.

test_data <- loadData()

}
