% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margalef.R
\name{margalef}
\alias{margalef}
\title{margalef - Menhinick's index calculating function}
\usage{
margalef(
  df,
  first.col = 2,
  plot = T,
  ylab = "Margalef's index value",
  xlab = "Samples",
  col = "gray",
  ...
)
}
\arguments{
\item{df}{xxx}

\item{first.col}{xxx}

\item{plot}{xxx}

\item{ylab}{xxx}

\item{xlab}{xxx}

\item{col}{xxx}

\item{...}{xxx}
}
\value{
xxx
}
\description{
margalef - Menhinick's index calculating function
}
\examples{
xxx

}
