% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menhinick.R
\name{menhinick}
\alias{menhinick}
\title{menhinick - Margalef’s index calculating function}
\usage{
menhinick(
  df,
  first.col = 2,
  plot = T,
  col = "gray",
  ylab = "Menhinick´s index value",
  xlab = "Samples",
  ...
)
}
\arguments{
\item{df}{xxx}

\item{first.col}{xxx}

\item{plot}{xxx}

\item{col}{xxx}

\item{ylab}{xxx}

\item{xlab}{xx}

\item{...}{xx}
}
\value{
xxx
}
\description{
menhinick - Margalef’s index calculating function
}
\examples{
xxx
}
