% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shannon.R
\name{shannon}
\alias{shannon}
\title{shannon - Shannon index and evenness}
\usage{
shannon(x, table = "CE", plot = T,  log = "e", ylab = "Shannon index value", xlab = "Samples", col = "black", ...)
}
\arguments{
\item{x}{A data frame uploaded by the loadData function (important because of the first column setting) containing the uploaded user dataset (list of taxa in the first column, followed by abundance columns with sample names in a header).}

\item{table}{Table output type. In case the table is equal to ‘C’ the table will be created only for Shannon calculations (Shannon index, the lower limit of Shannon index and the highest limit of Shannon index), for table equal to ‘E’ the output will be the table for evenness of Shannon index. In the case of table equal to ‘CE’ a sheet of two table type elements will be created for both Shannon calculations and evenness.}

\item{plot}{Should a plot for results of calculations be plotted? By default, the plot is rendered.}

\item{log}{Logarithm base type. The possibilities are ‘e’, for a base equal to Euler's number and ‘2’, for a base equal to two.}

\item{ylab}{The text for the y axis label in plot. By default, the text is setted as ‘Shannon index value’.}

\item{xlab}{The text for the x axis label. By default, the text is setted as ‘Samples’.}

\item{col}{The colour used in the plot. By default, it is set as ‘gray’.}

\item{...}{Arguments to be passed to methods, such graphical parameters. Many methods will acccept the following arguments.}
}
\value{
A data frame consisting of a column of calculated Shannon calculations (‘H´’, ‘Hmin’, ‘Hmax’) and evenness (‘Even1’, ‘Even2’) with samples in the rows.
}
\description{
Calculates Shannon index and evenness and displays the results of the calculation in graphical and tabular form.
}
\examples{
# Calculation of the Shannon calculations for the 'test_data' and plotted results. The base of the logaritm is set to 2'. Data frame of the results is stored in an object called 'shannon_tab'.

shannon_tab <- shannon(test_data, log = "2", table = "C")

}
