% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soren.eco.R
\name{soren}
\alias{soren}
\title{soren - Sørensen and Bray-Curtis similarity index calculating function}
\usage{
soren(df, first.col = 2, plot = T, cex = 0.75)
}
\arguments{
\item{df}{A data frame containing uploaded user dataset (list of taxa in first column, followed by columns of abundances with sample names in a header).}

\item{plot}{Should a plot for results of calculations be plotted? By default, the plot is rendered.}

\item{cex}{Numeric. Size of the text in the plot cells.}
}
\value{
xxx
}
\description{
Calculation and graphical display of the results of the Sørensen and Bray-Curtis index. similarity index
}
\examples{
soren(test_data)
}
